requirements.txt
flask
qrcode[pil]
pillow
gunicorn
requests
beautifulsoup4